refactor: follow hyphenated naming convention
