+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { createRoot } from 'react-dom/client'
 import { ReactQueryProvider } from 'react-query-provider'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
-import { GraphqlProvider } from 'graphqlRequestProvider'
+import { GraphqlProvider } from 'graphql-request-provider'
 import { ThemeProvider } from 'theme'
 
 import App from './App'
